+++ OnRail/Extensions/Try/TryExtensions.Async.cs
@@ -1,5 +1,3 @@
-using OnRail.ResultDetails;
-
 namespace OnRail.Extensions.Try;
 
 public static partial class TryExtensions {
@@ -7,7 +5,7 @@ public static partial class TryExtensions {
         Func<Task<T>> function,
         int numOfTry = 1
     ) {
-        var errors = new List<Exception>(numOfTry);
+        var errors = new List<object>(numOfTry);
 
         for (var counter = 0; counter < numOfTry; counter++) {
             try {
@@ -18,7 +16,8 @@ public static partial class TryExtensions {
             }
         }
 
-        return Result<T>.Fail(GenerateExceptionError(errors, numOfTry));
+        var errorDetail = TryHelper.GenerateError(errors, numOfTry);
+        return Result<T>.Fail(errorDetail);
     }
 
     //TODO: Test
@@ -52,8 +51,7 @@ public static partial class TryExtensions {
             }
         }
 
-        var errorDetail = new ErrorDetail(moreDetails: errors);
-        errorDetail.AddDetail(new {numOfTry});
+        var errorDetail = TryHelper.GenerateError(errors, numOfTry);
         return Result.Fail(errorDetail);
     }
 
@@ -85,8 +83,7 @@ public static partial class TryExtensions {
             }
         }
 
-        var errorDetail = new ErrorDetail(moreDetails: errors);
-        errorDetail.AddDetail(new {numOfTry});
+        var errorDetail = TryHelper.GenerateError(errors, numOfTry);
         return Result<T>.Fail(errorDetail);
     }
 
@@ -94,7 +91,7 @@ public static partial class TryExtensions {
         Func<Task> function,
         int numOfTry = 1
     ) {
-        var errors = new List<Exception>(numOfTry);
+        var errors = new List<object>(numOfTry);
 
         for (var counter = 0; counter < numOfTry; counter++) {
             try {
@@ -106,7 +103,8 @@ public static partial class TryExtensions {
             }
         }
 
-        return Result.Fail(GenerateExceptionError(errors, numOfTry));
+        var errorDetail = TryHelper.GenerateError(errors, numOfTry);
+        return Result.Fail(errorDetail);
     }
 
     public static async Task<Result> Try(
@@ -131,8 +129,7 @@ public static partial class TryExtensions {
             }
         }
 
-        var errorDetail = new ErrorDetail(moreDetails: errors);
-        errorDetail.AddDetail(new {numOfTry});
+        var errorDetail = TryHelper.GenerateError(errors, numOfTry);
         return Result.Fail(errorDetail);
     }
 
@@ -162,7 +159,7 @@ public static partial class TryExtensions {
         Action action,
         int numOfTry = 1
     ) {
-        var errors = new List<Exception>(numOfTry);
+        var errors = new List<object>(numOfTry);
 
         for (var counter = 0; counter < numOfTry; counter++) {
             try {
@@ -175,7 +172,8 @@ public static partial class TryExtensions {
             }
         }
 
-        return Result.Fail(GenerateExceptionError(errors, numOfTry));
+        var errorDetail = TryHelper.GenerateError(errors, numOfTry);
+        return Result.Fail(errorDetail);
     }
 
     public static async Task<Result> Try(
@@ -202,8 +200,7 @@ public static partial class TryExtensions {
             }
         }
 
-        var errorDetail = new ErrorDetail(moreDetails: errors);
-        errorDetail.AddDetail(new {numOfTry});
+        var errorDetail = TryHelper.GenerateError(errors, numOfTry);
         return Result.Fail(errorDetail);
     }
 
@@ -231,8 +228,7 @@ public static partial class TryExtensions {
             }
         }
 
-        var errorDetail = new ErrorDetail(moreDetails: errors);
-        errorDetail.AddDetail(new {numOfTry});
+        var errorDetail = TryHelper.GenerateError(errors, numOfTry);
         return Result.Fail(errorDetail);
     }